DiscordWebhook diff reporter backend oldal és feliratkozás logika
