+++ laravel-backend/app/Console/Kernel.php
@@ -13,6 +13,7 @@ class Kernel extends ConsoleKernel
     protected function schedule(Schedule $schedule): void
     {
         $schedule->command('jtq:scrape')->dailyAt('02:00');
+        $schedule->command('jtq:discord-webhook')->dailyAt('07:00');
     }
 
     /**
