+++ laravel-backend/app/Http/Requests/PublicApi/DiscordSubscribeRequest.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace App\Http\Requests\PublicApi;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class DiscordSubscribeRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     */
+    public function authorize(): bool
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array|string>
+     */
+    public function rules(): array
+    {
+        return [
+            'webhook_url' => 'url|required',
+        ];
+    }
+}
