+++ laravel-backend/app/Models/Subscription.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ * App\Models\Subscription.
+ *
+ * @property int                             $id
+ * @property \Illuminate\Support\Carbon|null $created_at
+ * @property \Illuminate\Support\Carbon|null $updated_at
+ * @property string                          $last_sent_at
+ * @property string                          $type
+ * @property mixed                           $settings
+ * @property int                             $enabled
+ *
+ * @method static \Illuminate\Database\Eloquent\Builder|Subscription newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|Subscription newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|Subscription query()
+ * @method static \Illuminate\Database\Eloquent\Builder|Subscription whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Subscription whereEnabled($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Subscription whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Subscription whereLastSentAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Subscription whereSettings($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Subscription whereType($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|Subscription whereUpdatedAt($value)
+ *
+ * @mixin \Eloquent
+ */
+class Subscription extends Model
+{
+    use HasFactory;
+
+    protected $fillable = [
+        'type', 'settings',
+    ];
+
+    protected $casts = [
+        'settings' => 'array',
+    ];
+}
